requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
beautifulsoup4==4.10.0
cached-property==1.5.2
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cryptography==36.0.1
fastapi==0.73.0
h11==0.13.0
httptools==0.3.0
idna==3.3
igraph==0.9.9
iniconfig==1.1.1
networkx==2.6.3
numpy==1.22.2
packaging==21.3
pandas==1.4.0
pdfminer.six==20211012
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
texttable==1.6.4
tomli==2.0.0
typing-extensions==4.0.1
uvicorn==0.17.4
uvloop==0.16.0
watchgod==0.7
websockets==10.1
whooshy==2.8.9
